+++ finagle-netty4/src/main/scala/com/twitter/finagle/netty4/exportNetty4MetricsAndRegistryEntries.scala
@@ -1,14 +1,10 @@
 package com.twitter.finagle.netty4
 
 import com.twitter.concurrent.Once
-import com.twitter.finagle.stats.{
-  FinagleStatsReceiver,
-  Gauge,
-  Verbosity,
-  VerbosityAdjustingStatsReceiver
-}
+import com.twitter.finagle.stats.{FinagleStatsReceiver, Gauge, Verbosity, VerbosityAdjustingStatsReceiver}
 import com.twitter.util.registry.GlobalRegistry
 import io.netty.buffer.{PoolArenaMetric, PooledByteBufAllocator}
+import io.netty.channel.epoll.Epoll
 import scala.collection.JavaConverters._
 import scala.collection.mutable
 
@@ -126,7 +122,7 @@ private object exportNetty4MetricsAndRegistryEntries {
 
     GlobalRegistry.get.put(
       Seq("library", "netty4", "native epoll enabled"),
-      nativeEpoll.enabled.toString
+      (useNativeEpoll() && Epoll.isAvailable).toString
     )
   }